requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3